Code for first prototype
It is a little rough, and a little blunt. A few times I felt like a more elegant solution was possible, I will continue to refine it. Overall a good basis for a mercury switch (AKA flip switch) pedometer. The code for the pedometer works well! The matrix part is not quite there. But that may have had more to do with me dropping it. We will see in future, more permanent iterations!